Service session picker records cached months (EDDBOOK-120)
Now that some sessions for a given month can overflow into the next,
the way to check if a month has been fetched and cached can no longer
be a simple cache existence check. A separate "registry" is maintained
that records which months have been requested, fetched and cached.

EDDBOOK-120 #time 20m